+++ Sources/SuperwallKit/Config/Models/Config.swift
@@ -79,7 +79,7 @@ struct Config: Codable, Equatable {
     case featureFlags = "toggles"
     case preloadingDisabled = "disablePreload"
     case attribution = "attributionOptions"
-    case products = "products"
+    case products = "productsV3"
     case web2appConfig
     case iosAppId
   }
@@ -102,11 +102,11 @@ struct Config: Codable, Equatable {
     locales = Set(localization.locales.map { $0.locale })
     requestId = try values.decodeIfPresent(String.self, forKey: .requestId)
 
-    let appStoreProductItems = try values.decodeIfPresent(
+    let products = try values.decodeIfPresent(
       [Throwable<Product>].self,
       forKey: .products
     ) ?? []
-    products = appStoreProductItems.compactMap { try? $0.result.get() }
+    self.products = products.compactMap { try? $0.result.get() }
   }
 
   func encode(to encoder: Encoder) throws {