+++ Sources/SuperwallKit/Models/SW Product/SWProductSubscriptionPeriod.swift
@@ -4,6 +4,7 @@
 //
 //  Created by Yusuf Tör on 02/03/2022.
 //
+// swiftlint:disable type_body_length
 
 import StoreKit
 
@@ -81,6 +82,38 @@ struct SWProductSubscriptionPeriod: Codable {
 
   @available(iOS 15.0, tvOS 15.0, watchOS 8.0, *)
   init(period: StoreKit.Product.SubscriptionPeriod, numberOfPeriods: Int) {
+    self.numberOfUnits = period.value * numberOfPeriods
+    switch period.unit {
+    case .day:
+      self.unit = .day
+    case .week:
+      self.unit = .week
+    case .month:
+      self.unit = .month
+    case .year:
+      self.unit = .year
+    @unknown default:
+      self.unit = .unknown
+    }
+  }
+
+  init(period: SubscriptionPeriod, numberOfPeriods: Int) {
+    self.numberOfUnits = period.value * numberOfPeriods
+    switch period.unit {
+    case .day:
+      self.unit = .day
+    case .week:
+      self.unit = .week
+    case .month:
+      self.unit = .month
+    case .year:
+      self.unit = .year
+    @unknown default:
+      self.unit = .unknown
+    }
+  }
+
+  init(period: StripeProductType.StripeSubscriptionPeriod, numberOfPeriods: Int) {
     self.numberOfUnits = period.value * numberOfPeriods
     switch period.unit {
     case .day: