+++ Sources/SuperwallKit/Paywall/View Controller/Web View/Message Handling/PaywallMessage.swift
@@ -17,7 +17,8 @@ import Foundation
 
 {
   "event_name": "open_url",
-  "url": "https://example.com"
+  "url": "https://example.com",
+  "browser_type": "payment_sheet" // optional
 }
 
 */
@@ -47,6 +48,7 @@ enum PaywallMessage: Decodable, Equatable {
   case restore
   case openUrl(_ url: URL)
   case openUrlInSafari(_ url: URL)
+  case openPaymentSheet(_ url: URL)
   case openDeepLink(url: URL)
   case purchase(productId: String)
   case custom(data: String)
@@ -91,7 +93,8 @@ enum PaywallMessage: Decodable, Equatable {
     case version
     case name
     case params
-    case reviewType = "reviewType"
+    case reviewType
+    case browserType
   }
 
   enum PaywallMessageError: Error {
@@ -120,7 +123,8 @@ enum PaywallMessage: Decodable, Equatable {
       case .openUrl:
         if let urlString = try? values.decode(String.self, forKey: .url),
           let url = URL(string: urlString) {
-          self = .openUrl(url)
+          let browserType = try? values.decode(String.self, forKey: .browserType)
+          self = browserType == "payment_sheet" ? .openPaymentSheet(url) : .openUrl(url)
           return
         }
       case .openUrlInSafari: